separated data control funcs into three files: manga, chapter, and misc to declutter each
