+++ components/homePageComps/latestChapters/latestChapters.jsx
@@ -1,4 +1,5 @@
-import { getChapterScansGroup, timeSince } from "@/utils/dataManipulation";
+import { getChapterScansGroup } from "@/utils/dataManipulation/chapter";
+import { timeSince } from "@/utils/dataManipulation/misc";
 import Image from "next/image";
 import Link from "next/link";
