+++ utils/dataManipulation/misc.js
@@ -1,0 +1,23 @@
+/* -- Misc functions -- */
+
+export const timeSince = (timestamp) => {
+    const now = new Date();
+    const past = new Date(timestamp);
+    const diffInSeconds = Math.floor((now - past) / 1000);
+
+    const minutes = Math.floor(diffInSeconds / 60);
+    if (minutes < 60) return `${minutes} minute${minutes !== 1 ? 's' : ''} ago`;
+
+    const hours = Math.floor(minutes / 60);
+    if (hours < 24) return `${hours} hour${hours !== 1 ? 's' : ''} ago`;
+
+    const days = Math.floor(hours / 24);
+    if (days < 7) return `${days} day${days !== 1 ? 's' : ''} ago`;
+
+    const weeks = Math.floor(days / 7);
+    if (weeks < 52) return `${weeks} week${weeks !== 1 ? 's' : ''} ago`;
+
+    const years = Math.floor(weeks / 52);
+    return `${years} year${years !== 1 ? 's' : ''} ago`;
+}
+
